+++ release/common/label.helper.js
@@ -14,4 +14,23 @@ export function formatLabel(label) {
     }
     return label;
 }
+/**
+ * Escapes a label.
+ *
+ * @export
+ * @param {*} label
+ * @returns {string}
+ */
+export function escapeLabel(label) {
+    return label.toLocaleString().replace(/[&'`"<>]/g, function (match) {
+        return {
+            '&': '&amp;',
+            '\'': '&#x27;',
+            '`': '&#x60;',
+            '"': '&quot;',
+            '<': '&lt;',
+            '>': '&gt;',
+        }[match];
+    });
+}
 //# sourceMappingURL=label.helper.js.map