+++ release/force-directed-graph/force-directed-graph.component.d.ts
@@ -52,4 +52,5 @@ export declare class ForceDirectedGraphComponent extends BaseChartComponent {
     onDragStart(node: any, $event: MouseEvent): void;
     onDrag($event: MouseEvent): void;
     onDragEnd($event: MouseEvent): void;
+    escape(label: any): string;
 }